fix: fixing dependencies, file arrays created
