+++ src/js/game.js
@@ -1,6 +1,5 @@
 // eslint-disable-next-line import/no-cycle
-import { arrayOfcards } from './script.js';
-// import {cardsGrid} from "./script.js";
+import { arrayOfcards } from './arrays.js';
 // eslint-disable-next-line import/no-cycle
 import { findWord, scoreArray } from './score.js';
 // eslint-disable-next-line import/no-cycle
