+++ src/js/score.js
@@ -1,8 +1,7 @@
 import cards from '../../data/cards.js';
 // eslint-disable-next-line import/no-cycle
-import {
-  arrayOfcards, BaseCard,
-} from './script.js';
+import { BaseCard } from './script.js';
+import { arrayOfcards } from './arrays.js';
 // eslint-disable-next-line import/no-cycle
 import checkToggle, { gameToggle } from './toggle.js';
 // eslint-disable-next-line import/no-cycle
